Start adding NoiseModule modifiers.
This commit introduces a new module (modifiers.rs) that contains a
few types for modifying the output of their constituent NoiseModules,
and also a trait for making them somewhat easier to produce. This
trait has also been implemented for all other NoiseModule types.

Secondly, this commit revamps the example1.rs file with the new API
format, and with a bit more explanation. It has also been passed an
optimization flag in the Makefile.

Finally, this commit improves the documentation of NoiseModule and
removes SeededModule (which was not in use anyway).

Signed-off-by: Atheriel <atheriel@gmail.com>